# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The WellNest Collection
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  name: string;
  category: string;
  location: string;
  nightlyPrice: number;
  guests: number;
  facilities: string[];
  contactEmail: string;
  images: string[];
  featured: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface CreatePropertyBody {
  name: string;
  category: string;
  location: string;
  nightlyPrice: number;
  guests: number;
  facilities: string[];
  contactEmail: string;
  images: string[];
  featured?: boolean;
}

export interface UpdatePropertyBody {
  name?: string;
  category?: string;
  location?: string;
  nightlyPrice?: number;
  guests?: number;
  facilities?: string[];
  contactEmail?: string;
  images?: string[];
  featured?: boolean;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListPropertiesParams = {
  category?: string;
  search?: string;
  featured?: string;
};