# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for The WellNest Collection
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: properties
    description: Property operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      description: Returns all published properties, optionally filtered by category or search
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"

  /properties/featured:
    get:
      operationId: getFeaturedProperties
      tags: [properties]
      summary: Get featured properties
      description: Returns featured properties for homepage showcase
      responses:
        "200":
          description: List of featured properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"

  /properties/categories:
    get:
      operationId: getPropertyCategories
      tags: [properties]
      summary: Get category stats
      description: Returns categories with property counts
      responses:
        "200":
          description: Categories with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"

  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property
      description: Returns a single property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Authenticate as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      description: Logout admin session
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Get admin session
      description: Returns current admin session status
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/properties:
    get:
      operationId: adminListProperties
      tags: [admin]
      summary: Admin list all properties
      description: Returns all properties for admin management
      responses:
        "200":
          description: List of all properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [admin]
      summary: Create a property
      description: Create a new property listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/properties/{id}:
    patch:
      operationId: updateProperty
      tags: [admin]
      summary: Update a property
      description: Update an existing property listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProperty
      tags: [admin]
      summary: Delete a property
      description: Delete a property listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Property deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Property:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        location:
          type: string
        nightlyPrice:
          type: number
        guests:
          type: integer
        facilities:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - location
        - nightlyPrice
        - guests
        - facilities
        - contactEmail
        - images
        - featured
        - createdAt
        - updatedAt

    CategoryStat:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count

    CreatePropertyBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        location:
          type: string
        nightlyPrice:
          type: number
        guests:
          type: integer
        facilities:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
      required:
        - name
        - category
        - location
        - nightlyPrice
        - guests
        - facilities
        - contactEmail
        - images

    UpdatePropertyBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        location:
          type: string
        nightlyPrice:
          type: number
        guests:
          type: integer
        facilities:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error