# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The WellNest Collection
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all published properties, optionally filtered by category or search
 * @summary List all properties
 */
export const ListPropertiesQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.string().optional(),
});

export const ListPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListPropertiesResponse = zod.array(ListPropertiesResponseItem);

/**
 * Returns featured properties for homepage showcase
 * @summary Get featured properties
 */
export const GetFeaturedPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetFeaturedPropertiesResponse = zod.array(
  GetFeaturedPropertiesResponseItem,
);

/**
 * Returns categories with property counts
 * @summary Get category stats
 */
export const GetPropertyCategoriesResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
});
export const GetPropertyCategoriesResponse = zod.array(
  GetPropertyCategoriesResponseItem,
);

/**
 * Returns a single property by ID
 * @summary Get a property
 */
export const GetPropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * Authenticate as admin
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * Returns current admin session status
 * @summary Get admin session
 */
export const AdminMeResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * Returns all properties for admin management
 * @summary Admin list all properties
 */
export const AdminListPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const AdminListPropertiesResponse = zod.array(
  AdminListPropertiesResponseItem,
);

/**
 * Create a new property listing
 * @summary Create a property
 */
export const CreatePropertyBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean().optional(),
});

/**
 * Update an existing property listing
 * @summary Update a property
 */
export const UpdatePropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePropertyBody = zod.object({
  name: zod.string().optional(),
  category: zod.string().optional(),
  location: zod.string().optional(),
  nightlyPrice: zod.number().optional(),
  guests: zod.number().optional(),
  facilities: zod.array(zod.string()).optional(),
  contactEmail: zod.string().optional(),
  images: zod.array(zod.string()).optional(),
  featured: zod.boolean().optional(),
});

export const UpdatePropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  location: zod.string(),
  nightlyPrice: zod.number(),
  guests: zod.number(),
  facilities: zod.array(zod.string()),
  contactEmail: zod.string(),
  images: zod.array(zod.string()),
  featured: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * Delete a property listing
 * @summary Delete a property
 */
export const DeletePropertyParams = zod.object({
  id: zod.coerce.number(),
});