# lib/api-zod/src/generated/types/adminLoginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The WellNest Collection
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginBody {
  password: string;
}