# lib/api-zod/src/generated/types/categoryStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The WellNest Collection
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryStat {
  category: string;
  count: number;
}