# lib/api-zod/src/generated/types/updatePropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The WellNest Collection
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePropertyBody {
  name?: string;
  category?: string;
  location?: string;
  nightlyPrice?: number;
  guests?: number;
  facilities?: string[];
  contactEmail?: string;
  images?: string[];
  featured?: boolean;
}